Extract intrinsics and data types into focused modules
Break up lib.rs and cpu.rs into single responsibility modules and
consolidate low-level architectural helpers:

- Add intrinsics.rs for x86_64 wrappers
- Extract comm_buffer.rs/request_target.rs/page_ownership.rs.
- Extract semaphore.rs out of cpu.rs.
- No functional changes
