+++ patina_mm_supervisor/src/supervisor_handlers/system_handlers.rs
@@ -12,7 +12,8 @@
 use r_efi::efi;
 
 use crate::{
-    is_buffer_inside_mmram, read_cr3,
+    intrinsics::read_cr3,
+    is_buffer_inside_mmram,
     state::{init_state, security_state},
 };
